🐎 upgrade function

+++ project/server/modules/crawl_pipeline.py
@@ -10,6 +10,10 @@ import subprocess
 import pandas as pd
 import importlib.util
 from datetime import datetime
+import scrapy
+
+from crawler.crawler import google_crawl
+
 
 class CrawlManager():
     def __init__(self,
@@ -23,8 +27,8 @@ class CrawlManager():
 
 
     def run(self, except_spider:list[str]=[]):
-        except_spider = ['PpomppuSpider'] + except_spider
-        
+        # except_spider = ['PpomppuSpider'] + except_spider
+
         container_id = self.run_docker_splash()
 
         if self.base_dir.is_dir() == False:
@@ -34,6 +38,7 @@ class CrawlManager():
         self.run_scrapy(spider_command)
         self.remove_docker_container(container_id)
         self.merge_csv_files()
+        # self.scrape_google_play()
 
 
     @staticmethod
@@ -65,12 +70,28 @@ class CrawlManager():
         scrapy_command = ""
         for spider in spiders:
             scrapy_command += f"scrapy crawl {spider} -a user_id={self.user_id} -a keyword={self.keyword} -o {self.base_dir}/A_{spider.lower().split('spider')[0]}.csv \n"
-
+        print(scrapy_command)
         return scrapy_command
 
 
-    def _get_spider_name(self, except_spider):
+    # def _get_spider_name(self, except_spider):
 
+    #     spiders_list = []
+
+    #     for modules in self.module_path.rglob('./*.py'):
+    #         if (modules.name == '__init__.py') or (modules.name == None):
+    #             continue
+
+    #         spec = importlib.util.spec_from_file_location(modules.name, self.module_path / modules.stem / modules.name)
+    #         module = importlib.util.module_from_spec(spec)
+    #         spec.loader.exec_module(module)
+
+    #         for name, obj in inspect.getmembers(module):
+    #             if inspect.isclass(obj) and ('Spider' in name):
+    #                 spiders_list.append(name)
+
+    #     return [item for item in spiders_list if item not in except_spider]
+    def _get_spider_name(self, except_spider):
         spiders_list = []
 
         for modules in self.module_path.rglob('./*.py'):
@@ -82,20 +103,21 @@ class CrawlManager():
             spec.loader.exec_module(module)
 
             for name, obj in inspect.getmembers(module):
-                if inspect.isclass(obj) and ('Spider' in name):
+                if inspect.isclass(obj) and 'Spider' in name and issubclass(obj, scrapy.Spider):
                     spiders_list.append(name)
-
+        print([item for item in spiders_list if item not in except_spider])
         return [item for item in spiders_list if item not in except_spider]
 
 
     def merge_csv_files(self):
         crawl_dir_list = list(self.base_dir.parent.iterdir())
-        print(crawl_dir_list)
 
         for crawl_dir in crawl_dir_list:
+            # csv_files = list(crawl_dir.glob('A_*.csv'))
+            csv_files = [file for file in crawl_dir.glob('A_*.csv') if file.is_file()]
 
-            csv_files = list(crawl_dir.glob('A_*.csv'))
-            # dataframes_list = [pd.read_csv(file) for file in csv_files if not pd.read_csv(file).empty()]
+            if not csv_files:
+                continue
 
             dataframes_list = []
             for file in csv_files:
@@ -103,18 +125,56 @@ class CrawlManager():
                     df = pd.read_csv(file)
                     if not df.empty:
                         dataframes_list.append(df)
-                
+
                 except pd.errors.EmptyDataError:
                     continue
 
-            merged_df = pd.concat(dataframes_list)
-            merged_df.to_csv(crawl_dir / 'merged_data.csv', index=False)
+            if not dataframes_list:
+                continue
 
-            for csv_file in csv_files:
-                os.remove(csv_file)
+            merged_df = pd.concat(dataframes_list)
+            print(f"{crawl_dir}에서 {len(dataframes_list)}개의 데이터프레임을 병합 중입니다.")
+
+            if not merged_df.empty:
+                merged_df.to_csv(crawl_dir / 'merged_data.csv', index=False)
+                print(f"병합된 데이터를 {crawl_dir / 'merged_data.csv'}에 저장했습니다.")
+
+                for csv_file in csv_files:
+                    os.remove(csv_file)
+            else:
+                print(f"병합된 데이터프레임이 비어 있어 {crawl_dir}에 CSV로 저장되지 않았습니다.")
+
+
+
+    # def merge_csv_files(self):
+    #     crawl_dir_list = list(self.base_dir.parent.iterdir())
+
+    #     for crawl_dir in crawl_dir_list:
+
+    #         csv_files = list(crawl_dir.glob('A_*.csv'))
+    #         # dataframes_list = [pd.read_csv(file) for file in csv_files if not pd.read_csv(file).empty()]
+
+    #         dataframes_list = []
+    #         for file in csv_files:
+    #             try:
+    #                 df = pd.read_csv(file)
+    #                 if not df.empty:
+    #                     dataframes_list.append(df)
+
+    #             except pd.errors.EmptyDataError:
+    #                 continue
+
+    #         merged_df = pd.concat(dataframes_list)
+    #         merged_df.to_csv(crawl_dir / 'merged_data.csv', index=False)
+
+    #         for csv_file in csv_files:
+    #             os.remove(csv_file)
+
 
 
 if __name__ == "__main__":
-    cm = CrawlManager('asdf1234', '레이니75')
+    cm = CrawlManager('asdf1234', '멜론')
     # print(cm.get_spider_command(except_spider=['PpomppuSpider']))
+    # cm.run(except_spider=['QuesarzoneSpider'])
     cm.run()
+
